TextField: add input props and additional box props
Adds the option to forward additional props to the input with
`inputProps`. This is especially useful for HTML attributes like
`autocomplete` and `autofocus` or `aria-*` annotatations.

You can also pass any property that applies to `<Box>` that will be
forwarded to the root element of the input field. Examples are `bg` or
`m` or `p`.