Add test for matching `emotion` versions
We need to ensure we install matching versions of Emotion
(`@emotion/styled` and `@emotion/core`) because if we don't we'd break
our styling entirely. This is because different versions of Emotion
operate on distinct contexts, and, thus, the `styled` function we export
from the design system doesn't have access to the exported  `theme`.

This adds a test that should(tm) be red every time we accidentally have
mismatched Emotion versions. It's basically testing if a component
created with `styled()` has access to the `theme`.