Fix missing key press handling on `TextLinkButton`
We use a `<div role="button"` for `TextLinkButton` so that we can use it
within a block of text because this makes it an inline element.

However, when doing that, we missed to add key press handlers for Space
and Enter, which are required per spec:
https://www.w3.org/TR/wai-aria-practices/examples/button/button.html

Basically, normal buttons get triggered by pressing Space or Enter when
they are focused and you TextLinkButton should do that, too. This PR
does exactly that.