Fix shadows of `DropdownMenu`
The `DropdownMenu.Content` component had a `box-shadow` applied.
However, the shadow was removed by a rule in our global CSS reset:

```
*:focus:not(:focus-visible) {
  outline: none !important;
  box-shadow: none !important;
  border: none !important;
}
```

This commit fixes that by using a pseudo-element for the shadow instead,
because pseudo-elements aren't targeted by the global rule.

Since this rule has a purpose[0], we don't want to get rid of it just
yet. If it continues to trouble us, we might reconsider removing it.
Removing it would require us to change *a lot* of our focus styles,
though, so I'm really hesitant to do it.

[0] https://www.tpgi.com/focus-visible-and-backwards-compatibility/