Stop using negative top margin on Stack
If space prop is 5, Stack component will overlap top components and won't allow to click them. Removed the margin top negative space and stop providing space to the first child stack Item.
This change should be backwards compatible since we don't change what the user expects or the API of the Stack component.
Also added a story to test it.