Migrate dropdown menu to use Reach instead of Radix (QC-2017)
The library component from Radix UI does not support IE11 which breaks
the app on IE11. In this PR we replace Radix with Reach, however we're
breaking some functionality with it, especially the `asChild` attribute
and having to remove the functionality of nested menus. The component is
only used in two well-known places, so we're going with a minor release.