Exclude node_modules from transitive dependencies in bundle.
Since the last changes due to an unknown reason some of the transitive
dependencies in the `node_modules` folder of dependencies were also
transpiled with Babel and bundled. However, this is not necessary since
they're going to be installed together with this package in the target
environment.

Hence we exclude all `node_modules` folders from inclusion in the final
bundle.